fix survey delete button
