+++ app/src/features/surveys/view/SurveyDetails.tsx
@@ -5,17 +5,11 @@ import Typography from '@material-ui/core/Typography';
 import makeStyles from '@material-ui/core/styles/makeStyles';
 import SurveyGeneralInformation from 'features/surveys/view/components/SurveyGeneralInformation';
 import SurveyProprietaryData from 'features/surveys/view/components/SurveyProprietaryData';
-import React, { useContext } from 'react';
+import React from 'react';
 import { IGetProjectForViewResponse } from 'interfaces/useProjectApi.interface';
 import { IGetSurveyForViewResponse } from 'interfaces/useSurveyApi.interface';
 import { IGetAllCodeSetsResponse } from 'interfaces/useCodesApi.interface';
 import SurveyStudyArea from './components/SurveyStudyArea';
-import Button from '@material-ui/core/Button';
-import Icon from '@mdi/react';
-import { mdiDelete } from '@mdi/js';
-import { DialogContext } from 'contexts/dialogContext';
-import { useBiohubApi } from 'hooks/useBioHubApi';
-import { useHistory } from 'react-router';
 
 export interface ISurveyDetailsProps {
   surveyForViewData: IGetSurveyForViewResponse;
@@ -45,57 +39,12 @@ const useStyles = makeStyles((theme: Theme) => ({
 const SurveyDetails: React.FC<ISurveyDetailsProps> = (props) => {
   const { surveyForViewData, codes, refresh, projectForViewData } = props;
 
-  const biohubApi = useBiohubApi();
-  const history = useHistory();
   const classes = useStyles();
-  const dialogContext = useContext(DialogContext);
-
-  const defaultYesNoDialogProps = {
-    dialogTitle: 'Delete Survey',
-    dialogText: 'Are you sure you want to delete this survey, its attachments and associated observations?',
-    open: false,
-    onClose: () => dialogContext.setYesNoDialog({ open: false }),
-    onNo: () => dialogContext.setYesNoDialog({ open: false }),
-    onYes: () => dialogContext.setYesNoDialog({ open: false })
-  };
-
-  const showDeleteSurveyDialog = () => {
-    dialogContext.setYesNoDialog({
-      ...defaultYesNoDialogProps,
-      open: true,
-      onYes: () => {
-        deleteSurvey();
-        dialogContext.setYesNoDialog({ open: false });
-      }
-    });
-  };
-
-  const deleteSurvey = async () => {
-    try {
-      const response = await biohubApi.survey.deleteSurvey(projectForViewData.id, surveyForViewData.survey_details.id);
-
-      if (!response) {
-        return;
-      }
-
-      history.push(`/projects/${projectForViewData.id}/surveys`);
-    } catch (error) {
-      return error;
-    }
-  };
 
   return (
     <>
-      <Box mb={5} display="flex" justifyContent="space-between">
+      <Box mb={5}>
         <Typography variant="h2">Survey Details</Typography>
-        <Button
-          variant="outlined"
-          color="primary"
-          data-testid="delete-survey-button"
-          startIcon={<Icon path={mdiDelete} size={1} />}
-          onClick={showDeleteSurveyDialog}>
-          Delete Survey
-        </Button>
       </Box>
 
       <Box component={Paper} p={4}>
